requirements.txt
pandas 
flask 
numpy
gunicorn
sklearn
pillow